Implement playlist update and deletion functionality; add updatePlaylist and deletePlaylist methods in PlaylistService and PlaylistServiceImpl; enhance PlaylistController with update and delete endpoints; introduce PlaylistNotFoundException and PlaylistAccessDeniedException for error handling

+++ Backend/core-service/src/main/java/com/rhythmix/coreservice/service/impl/ArtistServiceImpl.java
@@ -80,7 +80,10 @@ public class ArtistServiceImpl implements ArtistService {
     @Override
     @Transactional
     public void deleteArtist(UUID artistId) {
-        log.info("Deleting artist {}", artistId);
+        if (!artistRepository.existsById(artistId)) {
+            throw new ArtistNotFoundException("Artist not with id '" + artistId + "' not found");
+        }
         artistRepository.deleteById(artistId);
+        log.info("Deleting artist {}", artistId);
     }
 }
